# Various small performance enhancements when sending messages via JMAP