--- a/core/src/test/java/org/apache/james/core/UsernameTest.java
+++ b/core/src/test/java/org/apache/james/core/UsernameTest.java
@@ -154,6 +154,12 @@ class UsernameTest {
     }
 
     @Test
+    void fromUsernameShouldThrowWhenMultipleDomainDelimiterTogether() {
+        assertThatThrownBy(() -> Username.of("aa@@aa"))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
     void fromUsernameShouldThrowWhenEndsWithDomainDelimiter() {
         assertThatThrownBy(() -> Username.of("aa@"))
             .isInstanceOf(IllegalArgumentException.class);