--- a/server/protocols/jmap-draft/src/main/java/org/apache/james/jmap/draft/model/SetMessagesRequest.java
+++ b/server/protocols/jmap-draft/src/main/java/org/apache/james/jmap/draft/model/SetMessagesRequest.java
@@ -168,6 +168,10 @@ public class SetMessagesRequest implements JmapRequest {
         return Maps.transformValues(update, func -> func.apply(converter));
     }
 
+    public boolean hasUpdates() {
+        return !update.isEmpty();
+    }
+
     public List<MessageId> getDestroy() {
         return destroy;
     }