## Before

```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                      29532 (OK=29532  KO=0     )
> min response time                                      6 (OK=6      KO=-     )
> max response time                                   5137 (OK=5137   KO=-     )
> mean response time                                   344 (OK=344    KO=-     )
> std deviation                                        325 (OK=325    KO=-     )
> response time 50th percentile                        241 (OK=241    KO=-     )
> response time 75th percentile                        370 (OK=370    KO=-     )
> response time 95th percentile                        848 (OK=848    KO=-     )
> response time 99th percentile                       1777 (OK=1777   KO=-     )
> mean requests/sec                                 40.678 (OK=40.678 KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                         27913 ( 95%)
> 800 ms < t < 1200 ms                                 748 (  3%)
> t > 1200 ms                                          871 (  3%)
> failed                                                 0 (  0%)
================================================================================
```


## After

```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                      29790 (OK=29790  KO=0     )
> min response time                                      8 (OK=8      KO=-     )
> max response time                                   5403 (OK=5403   KO=-     )
> mean response time                                   316 (OK=316    KO=-     )
> std deviation                                        256 (OK=256    KO=-     )
> response time 50th percentile                        248 (OK=248    KO=-     )
> response time 75th percentile                        327 (OK=327    KO=-     )
> response time 95th percentile                        691 (OK=691    KO=-     )
> response time 99th percentile                       1363 (OK=1363   KO=-     )
> mean requests/sec                                 40.977 (OK=40.977 KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                         28893 ( 97%)
> 800 ms < t < 1200 ms                                 460 (  2%)
> t > 1200 ms                                          437 (  1%)
> failed                                                 0 (  0%)
================================================================================
```